Error handling for method resolve failure
Renamed TargetDestroyedErrorHandlingPolicy -> ErrorHandlingPolicy. The enum is now generic enough to be used for multiple potential error sources.
Failure to resolve the target method at runtime will consult this policy and if applicable, log an error or throw an exception of the new exception type ResolveMethodFailedException.
Changed InvocationTargetDestroyedException to derive from base class JPBaseException.
Changed the existing 'target destroyed' error policy to default to LogError if not set in the global config.